devtools/server/tests/mochitest/test_framerate_03.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1023018 - Tests whether or not the framerate actor can handle time ranges.
-->
<head>
  <meta charset="utf-8">
  <title>Framerate actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

window.onload = function() {
  const {require} = ChromeUtils.import("resource://devtools/shared/Loader.jsm", {});
  const {DebuggerClient} = require("devtools/shared/client/debugger-client");
  const {DebuggerServer} = require("devtools/server/main");
  const Services = require("Services");

  // Always log packets when running tests.
  Services.prefs.setBoolPref("devtools.debugger.log", true);
  SimpleTest.registerCleanupFunction(function() {
    Services.prefs.clearUserPref("devtools.debugger.log");
  });

  SimpleTest.waitForExplicitFinish();

  const {FramerateFront} = require("devtools/shared/fronts/framerate");
  const START_TICK = 2000;
  const STOP_TICK = 3000;
  const TOTAL_TIME = 5000;

  DebuggerServer.init();
  DebuggerServer.registerAllActors();

  const client = new DebuggerClient(DebuggerServer.connectPipe());
  client.connect().then(function onConnect() {
    client.listTabs().then(function onListTabs(response) {
      const form = response.tabs[response.selected];
      const front = FramerateFront(client, form);

      front.startRecording().then(() => {
        /* eslint-disable max-nested-callbacks */
        window.setTimeout(() => {
          front.stopRecording(START_TICK, STOP_TICK).then(rawData => {
            onRecordingStopped(front, rawData);
          });
        }, TOTAL_TIME);
        /* eslint-enable max-nested-callbacks */
      });
    });
  });

  function onRecordingStopped(front, rawData) {
    ok(rawData, "There should be a recording available.");

    ok(!rawData.find(e => e < START_TICK),
      "There should be no tick before 2000ms.");
    ok(!rawData.find(e => e > STOP_TICK),
      "There should be no tick after 3000ms.");

    for (const tick of rawData) {
      info("Testing tick: " + tick);
      is(typeof tick, "number", "All values should be numbers.");
    }

    client.close().then(() => {
      DebuggerServer.destroy();
      SimpleTest.finish();
    });
  }
};
</script>
</pre>
</body>
</html>
